Lib: Move generic Taquito wrapper code to a separate file
